PR Feedback: Move guard_call_logging to tracing.
