+++ guardrails/utils/telemetry_utils.py
@@ -7,7 +7,7 @@ from opentelemetry import context
 from opentelemetry.context import Context
 from opentelemetry.trace import StatusCode, Tracer
 
-from guardrails.guard_call_logging import TraceHandler
+from guardrails.call_tracing.guard_call_logging import TraceHandler
 from guardrails.stores.context import get_tracer as get_context_tracer
 from guardrails.stores.context import get_tracer_context
 from guardrails.utils.casting_utils import to_string
